pyproject.toml
[tool.poetry]
name = "eWalletv1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
uuid = "^1.30"
Flask-RESTful = "^0.3.9"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
